+++ parseq/gui/plotOptions.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 __author__ = "Konstantin Klementiev"
-__date__ = "15 Jul 2022"
+__date__ = "30 Oct 2025"
 # !!! SEE CODERULES.TXT !!!
 
 from collections import OrderedDict