+++ parseq/help/conf_doc.py
@@ -60,3 +60,6 @@ html_show_sphinx = False
 html_show_copyright = False
 
 html_scaled_image_link = False
+
+# with mathjax@4, :math: is broken in QWebEnginePage, so we use @3:
+mathjax_path = 'https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js'